requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
braceexpand==0.1.7
click==8.1.7
discord.py==2.4.0
frozenlist==1.4.1
idna==3.7
import_expression==1.1.5
jishaku==2.5.2
multidict==6.0.5
psutil==6.0.0
six==1.16.0
yarl==1.9.4
